requirements.txt
amqp==5.3.1
asgiref==3.11.1
billiard==4.2.4
celery==5.3.6
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.2
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
Django==5.2.13
django-rest-passwordreset==1.5.0
djangorestframework==3.14.0
idna==3.11
kombu==5.6.2
packaging==26.1
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
pytz==2026.1.post1
PyYAML==6.0.3
requests==2.31.0
six==1.17.0
sqlparse==0.5.5
tzdata==2026.1
ujson==5.9.0
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0
